pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django~=6.0",
    "django-encrypted-model-fields~=0.6.5",
    "django-filter~=25.1",
    "djangorestframework~=3.16",
    "djangorestframework-gis~=1.2",
    "psycopg~=3.3",
    "meshcore~=2.2",
    "channels~=4.2",
    "channels-redis~=4.2",
    "daphne~=4.1",
    "pyserial~=3.5",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "ruff~=0.14",
]

[tool.ruff]
target-version = "py314"
line-length = 127
indent-width = 4
exclude = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
ignore = ["DJ001", "DJ006", "EM101", "ANN401"]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade (modernize Python code)
    "EM",     # flake8-errmsg
    "DJ",     # flake8-django
    "ANN",    # flake8-annotations (type hints)
    "TCH",    # flake8-type-checking (TYPE_CHECKING imports)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.pyright]
include = ["metro", "api"]
exclude = ["**/node_modules", "**/__pycache__", "**/migrations"]
typeCheckingMode = "basic"
pythonVersion = "3.14"
pythonPlatform = "All"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true
reportDuplicateImport = true
reportOptionalMemberAccess = true
reportOptionalSubscript = true
reportPrivateImportUsage = false
reportGeneralTypeIssues = true
reportUntypedFunctionDecorator = false
reportUntypedClassDecorator = false
reportUntypedBaseClass = false
reportConstantRedefinition = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
venvPath = "."
venv = ".venv"

